Add TODO in response to intermittent error
This commit adds a todo comment to the endpoint list component to solve an intermittent error with the fetch endpoints method. Sometimes the user is able to click through to a different route by the time the fetch call finishes. In this case the list component has already been unmounted, but a stray setState call still gets called.