Add Response Header Input Component
This commit updates adds a new component to the frontend dashboard. This component handles editing/displaying response headers.

This component offers several props to help control it. An initial set of headers can be passed in, the display mode can change from readonly to editable (and vice-versa), and a change handler is available for status updates.

Styling for custom tables is also updated to make the component look/feel closer to the rest of the form component.

This commit does not add code to render the component. Those changes will be added in a follow up commit.