Support Response Headers
This commit completes the last updates  needed to support mock response headers. Both the Detail and New Components are updated to add a headers array to the initial state. The headers change handler is also a special case (this is now handled). Finally, the headers array is included in the POST and PUT requests, so the headers are persisted in on the server side.

This commit additionally adjusts the components to disable lints where applicable. The most common case is when certain console logs or alerts are expected and used. TODO comments have been added nearby statements that should be removed in the future.