Support Mocked Response Headers (#1)
* Add TODO in response to intermittent error

This commit adds a todo comment to the endpoint list component to solve an intermittent error with the fetch endpoints method. Sometimes the user is able to click through to a different route by the time the fetch call finishes. In this case the list component has already been unmounted, but a stray setState call still gets called.

* Support Mocked Response Headers on Server

This commit updates the server-side to support response headers from incoming requests. The mock middleware handler is also updated to add headers if needed.

ESLint no-console warnings have been disabled in several locations as these console logs are expected.

* Add Response Header Input Component

This commit updates adds a new component to the frontend dashboard. This component handles editing/displaying response headers.

This component offers several props to help control it. An initial set of headers can be passed in, the display mode can change from readonly to editable (and vice-versa), and a change handler is available for status updates.

Styling for custom tables is also updated to make the component look/feel closer to the rest of the form component.

This commit does not add code to render the component. Those changes will be added in a follow up commit.

* Render the Header Table Input inside the Endpoint Form

This commit updates the Endpoint Form Component to import and render the header table input. This component takes advantage of the props provided to handle the editing state, as well as respond to input changes.

* Support Response Headers

This commit completes the last updates  needed to support mock response headers. Both the Detail and New Components are updated to add a headers array to the initial state. The headers change handler is also a special case (this is now handled). Finally, the headers array is included in the POST and PUT requests, so the headers are persisted in on the server side.

This commit additionally adjusts the components to disable lints where applicable. The most common case is when certain console logs or alerts are expected and used. TODO comments have been added nearby statements that should be removed in the future.