Calculate MIME type before route gets run
This commit updates the endpoint store to perform some initial checks when adding/updating an endpoint. A default mime type is given when it is not given (plaintext). Further. a specific JSON check is added to explicitly set or remove the JSON mime type based on if it can be successfully parsed. More of these checks should be added and this should possibly move to another utility function