Finish Implementing Endpoint GraphQL Schema
This commit completes the rest of the graphQL Schema up to a working point. There are some validation checks (mostly non-null and some like type checking), but more checks are necessary (condition requirement of redirect, etc.). 

The mime type enum is updated to include JSON and the correct value for JS. An additional Http method was added to hold the method for the endpoint.

the API has been split up correctly between the endpoint query and endpoint mutation objects. These methods simply rely on the root object, which is the store given to us in the setup function

The main API setup function has been updated to add the GraphQL route. The GraphiQL interface is also enabled for testing (should be hidden around a dev env flag).