Add GraphQL API (#2)
* Save only the name and value of header data

The header data coming into the request can sometimes contain additional data that is not relevant in the server's context. Instead of holding onto that data, only capture the relevant values when creating/updating an endpoint.

* Add a Separate Store Data Structure

This commit updates the server to use a separate store data structure for handling data persistance. This allows multiple sources to see/use the data instead of having the endpoints be tied to a specific file.

* Start Adding GraphQL Schema

This commit adds graphQL dependencies to the project and starts defining the graphQL schema. The endpoint types are defined, and the query will be added next.

* Use MimeType and Method

This commit updates the mock endpoints middleware to take advantage of the mime type when sending the content (unless it's been overridden by a custom header).

The middleware is also updated to use the new endpoint structure, rather than relying on the deprecated endpoint structure.

* Calculate MIME type before route gets run

This commit updates the endpoint store to perform some initial checks when adding/updating an endpoint. A default mime type is given when it is not given (plaintext). Further. a specific JSON check is added to explicitly set or remove the JSON mime type based on if it can be successfully parsed. More of these checks should be added and this should possibly move to another utility function

* Finish Implementing Endpoint GraphQL Schema

This commit completes the rest of the graphQL Schema up to a working point. There are some validation checks (mostly non-null and some like type checking), but more checks are necessary (condition requirement of redirect, etc.). 

The mime type enum is updated to include JSON and the correct value for JS. An additional Http method was added to hold the method for the endpoint.

the API has been split up correctly between the endpoint query and endpoint mutation objects. These methods simply rely on the root object, which is the store given to us in the setup function

The main API setup function has been updated to add the GraphQL route. The GraphiQL interface is also enabled for testing (should be hidden around a dev env flag).

* Update Frontend to use graphql API

This commit updates the frontend to use the new graphql API for interacting with endpoints. The old API can now be removed.

* Add nodemon config

This commit updates the project to add config options for nodemon. The most notable options is preventing updates to the `web` directory from causing a restart of the app server.

* Add helper function for graphQL Requests

This commit refactors out all shared parts of the graphQL requests to a separate helper function.

* Add Missing DeleteAll API Method

This commit adds a new mutation for deleting all the endpoints. This was previously missing from the graphQL API, but it has been added and is now being used by the frontend.

* Remove Old Endpoints API

This commit removes the old endpoints api since it is no longer in use. The graphQL api has been moved to the root route for endpoints.

* Enable GraphiQL only in Dev Mode