Add bin script
This commit updates the main.js file to export a start function instead of start the server automatically. The exported function is then used in the bin script to start the server. The package.json was updated to use the bin script instead of using the main app.js file.